Moved active monitoring endpoints to am module and improved rest
