+++ scripts/lua/rest/v2/get/ping/interfaces.lua
@@ -9,6 +9,7 @@ package.path = dirs.installdir .. "/scripts/lua/modules/?.lua;" .. package.path
 require "lua_utils"
 local json = require("dkjson")
 local rest_utils = require("rest_utils")
+local interface_utils = require("interface_utils")
 
 --
 -- Return all interfaces enabled by ContinuousPing
@@ -18,26 +19,6 @@ local rest_utils = require("rest_utils")
 --
 
 local rc = rest_utils.consts.success.ok
-local res = {}
-
-local interfaces = ntop.getPingIfNames()
-
-for id, ifname in pairs(interfaces) do
-  local custom_name = getHumanReadableInterfaceName(ifname)
-
-  if isEmptyString(custom_name) then
-    custom_name = nil
-  end
-
-  -- Note: returning in a format compatible with /lua/rest/v2/get/ntopng/interfaces.lua
-  res[#res + 1] = {
-    ifid = tonumber(id),
-    ifname = ifname,
-    name = custom_name or ifname,
-    is_packet_interface = true,
-    is_pcap_interface = false,
-    is_zmq_interface = false,
-  }
-end
+local res = interface_utils.get_pingable_interfaces()
 
 rest_utils.answer(rc, res)
